Move heap objects into the HeapObject class
