+++ csharp/src/SeedLang/Ast/Function.cs
@@ -12,13 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System.Collections.Generic;
 using SeedLang.Runtime;
 
 namespace SeedLang.Ast {
   // A function value type that is only used in SeedAst component to encapsulate AST function
   // declearation statement.
-  internal class Function : IFunction {
+  internal class Function : HeapObject.IFunction {
     private readonly FuncDefStatement _funcDef;
     private readonly Executor _executor;
