--- csharp/src/SeedLang/Runtime/IFunction.cs
@@ -1,24 +1,0 @@
-// Copyright 2021-2022 The SeedV Lab.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-using System.Collections.Generic;
-
-namespace SeedLang.Runtime {
-  // The function interface of function value types.
-  internal interface IFunction {
-    // Calls the function with given arguments that locate in the "args" array starting from
-    // "offset". The number of arguments is "length".
-    Value Call(Value[] args, int offset, int length);
-  }
-}
