+++ csharp/tests/SeedLang.Tests/Runtime/HeapObjectsTests.cs
@@ -15,10 +15,12 @@
 using Xunit;
 
 namespace SeedLang.Runtime.Tests {
-  public class NumberRangeTests {
+  using Range = HeapObject.Range;
+
+  public class HeapObjectsTests {
     [Fact]
-    public void TestNumberRange() {
-      var range = new NumberRange(10);
+    public void TestRange() {
+      var range = new Range(10);
       Assert.Equal(10, range.Length);
       for (int i = 0; i < 10; i++) {
         Assert.Equal(i, range[i].AsNumber());
@@ -28,7 +30,7 @@ namespace SeedLang.Runtime.Tests {
       int stop = 10;
       int step = 2;
       int length = 4;
-      range = new NumberRange(start, stop, step);
+      range = new Range(start, stop, step);
       Assert.Equal(length, range.Length);
       for (int i = 0; i < length; i++) {
         Assert.Equal(start + i * step, range[i].AsNumber());
@@ -38,7 +40,7 @@ namespace SeedLang.Runtime.Tests {
       stop = 2;
       step = -3;
       length = 3;
-      range = new NumberRange(start, stop, step);
+      range = new Range(start, stop, step);
       Assert.Equal(length, range.Length);
       for (int i = 0; i < length; i++) {
         Assert.Equal(start + i * step, range[i].AsNumber());
@@ -48,7 +50,7 @@ namespace SeedLang.Runtime.Tests {
       stop = 10;
       step = -1;
       length = 0;
-      range = new NumberRange(start, stop, step);
+      range = new Range(start, stop, step);
       Assert.Equal(length, range.Length);
     }
   }
